+++ src/main/java/com/github/legoatoom/enderbench/block/entity/EnderBenchEntity.java
@@ -25,6 +25,7 @@ import com.github.legoatoom.enderbench.screen.EnderBenchScreenHandler;
 import io.netty.buffer.Unpooled;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.fabricmc.fabric.api.block.entity.BlockEntityClientSerializable;
 import net.fabricmc.fabric.api.network.ServerSidePacketRegistry;
 import net.fabricmc.fabric.api.server.PlayerStream;
 import net.minecraft.block.BlockState;
@@ -40,6 +41,7 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.screen.NamedScreenHandlerFactory;
 import net.minecraft.screen.ScreenHandler;
+import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.Tickable;
@@ -48,10 +50,11 @@ import net.minecraft.util.math.BlockPos;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Stream;
 
-public class EnderBenchEntity extends BlockEntity implements NamedScreenHandlerFactory, Inventory, Tickable {
+public class EnderBenchEntity extends BlockEntity implements NamedScreenHandlerFactory, Inventory, Tickable, BlockEntityClientSerializable {
 
     public static final int invSize = ModConfigs.EnderBenchSize;
     public static final double range = ModConfigs.EnderBenchRange;
@@ -158,11 +161,23 @@ public class EnderBenchEntity extends BlockEntity implements NamedScreenHandlerF
     }
 
     @Override
+    public void fromClientTag(CompoundTag tag) {
+        if (tag.contains("connection")) this.masterUuid = tag.getUuid("connection");
+    }
+
+    @Override
+    public CompoundTag toClientTag(CompoundTag tag) {
+        if (masterUuid != null) tag.putUuid("connection", masterUuid);
+        return tag;
+    }
+
+    @Override
     public void tick() {
         if (this.world == null) return;
         boolean active = false;
         isLocked = world.getBlockState(this.getPos()).get(EnderBenchBlock.LOCKED);
         if (!isLocked && this.world.isClient()){
+            masterUuid = null;
             // If it is not locked then everyone in range should be able to open it.
             // So this can be done in the client.
             List<? extends PlayerEntity> players = this.world.getPlayers();
@@ -187,25 +202,44 @@ public class EnderBenchEntity extends BlockEntity implements NamedScreenHandlerF
                     }
                 }
             }
-        } else if (isLocked && !this.world.isClient()) {
+            BlockState state = this.getCachedState();
+            world.setBlockState(pos, state.with(EnderBenchBlock.ACTIVE, active));
+        } else if (!this.world.isClient()) {
             // If it is locked then only the currently closes player can access it.
-            if (masterUuid == null){
-                PlayerEntity playerEntity = this.world.getClosestPlayer((double)this.pos.getX() + 0.5D,
-                        (double)this.pos.getY()  + 0.5D, (double)this.pos.getZ()  + 0.5D, range, false);
-                if (playerEntity != null) {
-                    active = true;
-                    Stream<PlayerEntity> stream = PlayerStream.around(this.world, this.pos, range * range);
-                    PacketByteBuf passedData = new PacketByteBuf(Unpooled.buffer());
-                    passedData.writeUuid(playerEntity.getUuid());
-                    passedData.writeBlockPos(this.pos);
+            if (isLocked) {
+                if (masterUuid == null) {
+                    PlayerEntity playerEntity = this.world.getClosestPlayer((double) this.pos.getX() + 0.5D,
+                            (double) this.pos.getY() + 0.5D, (double) this.pos.getZ() + 0.5D, range, false);
+                    if (playerEntity != null) {
+                        active = true;
+                        Stream<PlayerEntity> stream = PlayerStream.around(this.world, this.pos, range * range);
+                        PacketByteBuf passedData = new PacketByteBuf(Unpooled.buffer());
+                        masterUuid = playerEntity.getUuid();
+                        passedData.writeUuid(playerEntity.getUuid());
+                        passedData.writeBlockPos(this.pos);
 
-                    stream.forEach(player -> ServerSidePacketRegistry.INSTANCE.sendToPlayer(player,
-                            PacketIDs.S2C_SETCONNECTION_PACKET_ID, passedData));
+                        stream.forEach(player -> ServerSidePacketRegistry.INSTANCE.sendToPlayer(player,
+                                PacketIDs.S2C_SETCONNECTION_PACKET_ID, passedData));
+                    }
+                } else {
+                    active = true;
+                }
+            } else {
+                masterUuid = null;
+            }
+            BlockState state = this.getCachedState();
+            world.setBlockState(pos, state.with(EnderBenchBlock.ACTIVE, active));
+        } else if (isLocked && this.world.isClient){
+            if (masterUuid != null) {
+                IClientPlayerEntity a = ((IClientPlayerEntity) (world.getPlayerByUuid(masterUuid)));
+                if (a != null) {
+                    if (a.enderbench_getConnectedBenchPos() == null) {
+                        a.enderbench_setConnectedBenchPos(this.pos);
+                    }
                 }
             }
         }
-        BlockState state = this.getCachedState();
-        world.setBlockState(pos, state.with(EnderBenchBlock.ACTIVE, active));
+
     }
 
 
@@ -218,5 +252,6 @@ public class EnderBenchEntity extends BlockEntity implements NamedScreenHandlerF
         double distance = playerEntity.squaredDistanceTo(this.pos.getX(), this.pos.getY(), this.pos.getZ());
         return distance <= range * range;
     }
+
 
 }