+++ changelog.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Time is following the [Holocene Calendar](https://en.wikipedia.org/wiki/Holocene_calendar).
 
+## [1.0.1] - UNRELEASED
+### Optimized
+- Ender Bench Network Optimization.
+
+### Added
+- Config file for the range.s
+
+
 ## [0.1.0] - 12020-10-14
 ### Added
 - _Ender Bench_ Block with all its functionalities.